Creating methods for adding books to new created users. Also assign the new created users to the new created household - removing unneeded imports - changing long to int type - creating private post method in order to avoid code duplication

+++ 05. Web Services  & Integration Tests/BookRequests/Model/Household.cs
@@ -1,15 +1,12 @@
 ﻿using System;
-
-using System.Globalization;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
 
 namespace BookRequests.Model
 {
     public class Household
     {
         [JsonProperty("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [JsonProperty("name")]
         public string Name { get; set; }
